tests/wpt/web-platform-tests/css/css-transforms/parsing/translate-parsing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: parsing translate with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-translate">
<meta name="assert" content="translate supports the full grammar 'none | <length-percentage> [ <length-percentage> <length>? ]?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("translate", "none");

test_valid_value("translate", "0px");
test_valid_value("translate", "100%");

test_valid_value("translate", "100px 200%");
test_valid_value("translate", "100% 200px");

test_valid_value("translate", "100px 200px 300px");
test_valid_value("translate", "100% 200% 300px");

test_valid_value("translate", "calc(10px + 10%) calc(20px + 20%) calc(30px + 30em)");

test_valid_value("translate", "0", "0px");
test_valid_value("translate", "1px 2px 0", "1px 2px 0px");
</script>
</body>
</html>
